fix: Rewrite createSAMRequest function with template.
